move get_time_indexed_cuid into its own file
